refactor LMLogits
