+++ libai/layers/__init__.py
@@ -21,6 +21,7 @@ from .mlp import MLP
 from .layer_norm import LayerNorm
 from .transformer_layer import TransformerLayer
 from .cross_entropy import ParallelCrossEntropyLoss
+from .lm_logits import LMLogits
 
 
 __all__ = [
@@ -35,4 +36,5 @@ __all__ = [
     "LayerNorm",
     "TransformerLayer",
     "ParallelCrossEntropyLoss",
+    "LMLogits",
 ]
