+++ libai/models/bert_model.py
@@ -23,6 +23,7 @@ from libai.layers import (
     Linear,
     TransformerLayer,
     ParallelCrossEntropyLoss,
+    LMLogits,
 )
 from libai.utils import distributed as dist
 from oneflow import nn
@@ -200,50 +201,6 @@ class BertPreTrainingHeads(nn.Module):
         prediction_scores = self.predictions(sequence_output)
         seq_relationship_score = self.seq_relationship(pooled_output)
         return prediction_scores, seq_relationship_score
-
-
-class LMLogits(nn.Module):
-    def __init__(self, vocab_size, bias=False):
-        super().__init__()
-        self.bias = (
-            nn.Parameter(
-                flow.empty(
-                    (vocab_size,),
-                    dtype=flow.float32,
-                    placement=dist.get_layer_placement(-1),
-                    sbp=dist.get_nd_sbp([flow.sbp.broadcast, flow.sbp.split(0)]),
-                )
-            )
-            if bias
-            else None
-        )
-        if self.bias:
-            nn.init.zeros_(self.bias)
-
-    def forward(self, input, word_embeddings):
-        """LM logits using word embedding weights """
-        # input with sbp sign [S(0), B] and word_embeddings with sbp sign [S(0), B]
-
-        # NOTE(l1aoxingyu): This is for pipeline parallelism
-        # change word embedding placement from stage(0) to stage(-1)
-        w = word_embeddings.to_consistent(placement=input.placement)
-
-        # NOTE(l1aoxingyu): input x embed^T = logits with sbp sign
-        # [S(0), B] x [B, S(1)] --> [S(0), S(1)]
-        #     ↑          ↑               ↑
-        #   input      embed^T         logits
-        # Backward pass input.grad = logits.grad x embed with sbp sign
-        # [S(0), S(1)] x [B, S(0)] --> [S(0), P]
-        #     ↑             ↑               ↑
-        #  logits.grad    embed        input.grad
-        # When use input.grad as head node for backward pass, need to convert
-        # its sbp sign fromm [S(0), P] --> [S(0), B]
-        input = input.to_consistent(grad_sbp=input.sbp)
-
-        logits = flow._C.matmul(input, w, transpose_b=True)
-        if self.bias:
-            logits = logits + self.bias
-        return logits
 
 
 class BertLoss(nn.Module):
